add rebar3 support
make the buildpack support both rebar and rebar3.  The buildpack uses
the presence or absence of rebar, rebar3, or rebar.lock to pick which
build tool to use.  If none of these files are present, it will default
to using rebar.  Either rebar3 or rebar.lock or both will make the
buildpack use rebar3.